requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-filter==23.2
django-import-export==3.2.0
django-jalali-date==1.0.2
django-session-timeout==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jdatetime==4.1.1
MarkupPy==1.14
msgpack==1.0.5
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
reportlab==4.0.4
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
